backend/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
anthropic
python-dotenv
pytest
pytest-asyncio
httpx
apscheduler>=3.10.0
resend>=0.7.0